txscript: Separate short form script parsing.
This separates the test code related to parsing scripts using the
human-readable short form used throughout the tests into a separate file
so it can more easily be copied into other packages where it might be
useful for the tests as well.  Allowing it to be copied is preferred to
making it part of the public API of some package since it is an ad-hoc
format and it is not possible to access test-specific code across
package boundaries without making it part of the actual public API.
